+++ src/apps/weblib/js-api/errors.js
@@ -1,0 +1,6 @@
+export class CancelError extends Error {
+  constructor() {
+    super('animation canceled')
+    this.name = 'CancelError'
+  }
+}