+++ src/apps/weblib/js-api/objregistry.js
@@ -1,4 +1,6 @@
-export default class ObjectRegistry {
+export class CObject {}
+
+export class ObjectRegistry {
   constructor(fnFree) {
     this.finalizationRegistry = new FinalizationRegistry((rawCPointer) => {
       fnFree(rawCPointer)