+++ src/apps/weblib/js-api/render.js
@@ -1,4 +1,17 @@
-export default class Render {
+export class Rendering {
+  meta = { name: 'rendering' }
+
+  register(chart) {
+    this.chart = chart
+  }
+
+  enable(enabled) {
+    this.chart._validateModule()
+    this.chart.render.enabled = enabled
+  }
+}
+
+export class Render {
   init(ccanvas, update, canvas, log) {
     this.ccanvas = ccanvas
     this.enabled = true
@@ -7,9 +20,6 @@ export default class Render {
     this.offscreenContext = this.offscreenCanvas.getContext('2d')
     this.update = update
     this.mainCanvas = canvas
-    this.mainCanvas.font = 'Roboto Condensed'
-    this.mainCanvas.textAlign = 'left'
-    this.mainCanvas.textBaseline = 'top'
     this.context = canvas.getContext('2d')
     this.log = log
     this.updateCanvasSize()