+++ test/integration/test_cases/web_content/cookbook/rendering/dropshadow_on_marker.mjs
@@ -1,17 +1,14 @@
 import { data } from '../../../../test_data/chart_types_eu.mjs'
+import MarkerDropshadow from '../../../../utils/vizzu-markerdropshadow.mjs'
 
 const testSteps = [
-  (chart) => {
-    const setDropshadow = (event) => {
-      const ctx = event.renderingContext
-      ctx.shadowColor = '#00000060'
-      ctx.shadowBlur = 9
-      ctx.shadowOffsetX = 3
-      ctx.shadowOffsetY = 3
-    }
+  async (chart) => {
+    await import('https://unpkg.com/tinycolor2@1.6.0/dist/tinycolor-min.js')
+    return chart
+  },
 
-    chart.on('plot-marker-draw', setDropshadow)
-    chart.on('plot-marker-label-draw', setDropshadow)
+  (chart) => {
+    chart.feature(new MarkerDropshadow())
 
     return chart.animate({
       data,
@@ -21,9 +18,36 @@ const testSteps = [
         color: 'Joy factors',
         label: 'Value 2 (+)',
         title: 'Drop-shadow on the markers'
+      },
+      style: {
+        plot: {
+          marker: {
+            shadowColor: '#60000060',
+            shadowBlur: 9,
+            shadowOffsetX: 3,
+            shadowOffsetY: 3
+          }
+        }
       }
     })
-  }
+  },
+  (chart) =>
+    chart.animate({
+      data,
+      config: {
+        title: 'Smaller drop-shadow on the markers'
+      },
+      style: {
+        plot: {
+          marker: {
+            shadowColor: '#00006060',
+            shadowBlur: 5,
+            shadowOffsetX: -2,
+            shadowOffsetY: 2
+          }
+        }
+      }
+    })
 ]
 
 export default testSteps